move insight series resolver creation in to chain of responsibility
