+++ enterprise/internal/insights/query/capture_group_executor.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/grafana/regexp"
@@ -12,7 +11,6 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/compression"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/timeseries"
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/types"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -22,19 +20,18 @@ import (
 )
 
 type CaptureGroupExecutor struct {
-	db        database.DB
-	repoStore database.RepoStore
-	filter    compression.DataFrameFilter
-	clock     func() time.Time
+	justInTimeExecutor
 }
 
-func NewCaptureGroupExecutor(postgres, insightsDb dbutil.DB, clock func() time.Time) *CaptureGroupExecutor {
+func NewCaptureGroupExecutor(postgres, insightsDb dbutil.DB, clock func() time.Time) JITExecutor {
 	return &CaptureGroupExecutor{
-		db:        database.NewDB(postgres),
-		repoStore: database.Repos(postgres),
-		// filter:    compression.NewHistoricalFilter(true, clock().Add(time.Hour*24*365*-1), insightsDb),
-		filter: &compression.NoopFilter{},
-		clock:  clock,
+		justInTimeExecutor: justInTimeExecutor{
+			db:        database.NewDB(postgres),
+			repoStore: database.Repos(postgres),
+			// filter:    compression.NewHistoricalFilter(true, clock().Add(time.Hour*24*365*-1), insightsDb),
+			filter: &compression.NoopFilter{},
+			clock:  clock,
+		},
 	}
 }
 
@@ -61,17 +58,6 @@ func (c *CaptureGroupExecutor) Execute(ctx context.Context, query string, reposi
 		}
 		// uncompressed plan for now, because there is some complication between the way compressed plans are generated and needing to resolve revhashes
 		plan := c.filter.FilterFrames(ctx, frames, repoIds[repository])
-
-		generateTimes := func() map[time.Time]int {
-			times := make(map[time.Time]int)
-			for _, execution := range plan.Executions {
-				times[execution.RecordingTime] = 0
-				for _, recording := range execution.SharedRecordings {
-					times[recording] = 0
-				}
-			}
-			return times
-		}
 
 		// we need to perform the pivot from time -> {label, count} to label -> {time, count}
 		for _, execution := range plan.Executions {
@@ -109,7 +95,7 @@ func (c *CaptureGroupExecutor) Execute(ctx context.Context, query string, reposi
 			for _, timeGroupElement := range grouped {
 				value := timeGroupElement.Value
 				if _, ok := pivoted[value]; !ok {
-					pivoted[value] = generateTimes()
+					pivoted[value] = generateTimes(plan)
 				}
 				pivoted[value][execution.RecordingTime] = timeGroupElement.Count
 				for _, children := range execution.SharedRecordings {
@@ -143,43 +129,4 @@ func (c *CaptureGroupExecutor) Execute(ctx context.Context, query string, reposi
 		seriesCount++
 	}
 	return calculated, nil
-}
-
-func withCountUnlimited(s string) string {
-	if strings.Contains(s, "count:") {
-		return s
-	}
-	return s + " count:all"
-}
-
-func BuildFrames(numPoints int, interval timeseries.TimeInterval, now time.Time) []types.Frame {
-	current := now
-	times := make([]time.Time, 0, numPoints)
-	times = append(times, now)
-	times = append(times, now) // looks weird but is so we can get a frame that is the current point
-
-	for i := 0 - numPoints + 1; i < 0; i++ {
-		current = interval.StepBackwards(current)
-		times = append(times, current)
-	}
-
-	frames := make([]types.Frame, 0, len(times)-1)
-	for i := 1; i < len(times); i++ {
-		prev := times[i-1]
-		frames = append(frames, types.Frame{
-			From: times[i],
-			To:   prev,
-		})
-	}
-
-	sort.Slice(frames, func(i, j int) bool {
-		return frames[i].From.Before(frames[j].From)
-	})
-	return frames
-}
-
-type GeneratedTimeSeries struct {
-	Label    string
-	Points   []TimeDataPoint
-	SeriesId string
 }
