+++ enterprise/internal/insights/query/just_in_time_executor_test.go
